+++ applications/MappingApplication/tests/basic_mapper_tests.py
@@ -294,29 +294,6 @@ class BasicMapperTests(mapper_test_case.MapperTestCase):
         self.assertAlmostEqual(sum_origin, sum_destination)
 
 
-    def test_Is_conforming(self):
-        is_conforming = self.mapper.AreMeshesConforming()
-        self.assertTrue(is_conforming)
-
-    def test_Is_not_conforming(self):
-        non_conform_parameters = self.mapper_parameters.Clone()
-        non_conform_parameters.AddEmptyValue("search_radius").SetDouble(1e-6)
-
-        if data_comm.IsDistributed():
-            map_creator = MappingMPIExtension.MPIMapperFactory.CreateMapper
-        else:
-            map_creator = KratosMapping.MapperFactory.CreateMapper
-
-        non_conform_mapper = map_creator(
-            self.model_part_origin,
-            self.model_part_destination,
-            non_conform_parameters
-        )
-
-        is_conforming = non_conform_mapper.AreMeshesConforming()
-        self.assertFalse(is_conforming)
-
-
     # def test_UpdateInterface(self):
     #     pass
